build: enforce -Werror in debug configs, document quiet build/test recipes
Debug and debug-asan presets now fail on any warning (gated on
CONFIG:DEBUG since NDEBUG release builds turn assert-only locals
into false-positive -Wunused-variable). Fixes the one warning that
was live in the debug build here, a signed/unsigned compare in
VectorHandler. Also adds AGENTS.md notes for keeping build/test
output out of agent context (gtest_brief, redirect-and-tail-on-failure).

+++ src/solux/index/handler/VectorHandler.h
@@ -149,7 +149,7 @@ private:
     size_t start = scratch.size();
     scratch.resize(start + (size_t)vec.v.size());
     float invNorm = (float)(1.0 / std::sqrt(normSq));
-    for (int32_t i = 0; i < vec.v.size(); i++) {
+    for (int32_t i = 0; i < (int32_t)vec.v.size(); i++) {
       scratch[start + (size_t)i] = vec.v[i] * invNorm;
     }
     return bytesOf(scratch.data() + start, vec.v.size());
